MUC Cleaner - Refactored, removed renderer from controllers.
